+++ src/main.rs
@@ -135,6 +135,30 @@ struct RecentWorkspace {
     url: String,
 }
 
+/// Compute the score of matching `workspace` against `terms`.
+///
+/// If all terms match the name each term contributes a score of 10; this makes sure
+/// that precise matches in the name boost the score somewhat to the top.
+///
+/// If all terms match the URL each term contributes 1 to score, scaled by the relative position
+/// of the right-most match, assuming that URL paths typically go from least to most specific segment,
+/// to the farther to the right a term matches the more specific it was.
+fn match_score<S: AsRef<str>>(workspace: &RecentWorkspace, terms: &[S]) -> f64 {
+    let name = workspace.name.to_lowercase();
+    let path = workspace.url.to_lowercase();
+    let name_score = terms.iter().try_fold(0.0, |score, term| {
+        name.contains(&term.as_ref().to_lowercase())
+            .then(|| score + 10.0)
+            .ok_or(())
+    });
+    let path_score = terms.iter().try_fold(0.0, |score, term| {
+        path.rfind(&term.as_ref().to_lowercase())
+            .ok_or(())
+            .map(|index| score + 1.0 * (index as f64 / path.len() as f64))
+    });
+    name_score.unwrap_or_default() + path_score.unwrap_or_default()
+}
+
 /// Find all workspaces from `workspaces` which match the given `terms`.
 ///
 /// `workspaces` is an iterator over pairs of `(id, workspace)`.
@@ -142,35 +166,31 @@ struct RecentWorkspace {
 /// For each `workspace` match `terms` against the name and the `url` and return
 /// a vector with all `id`s of worksapces which match.
 ///
-/// Currently this simply checks whether all terms are container in `url` or `name`
-/// but this is subject to change.
+/// For each workspace compute the score with `match_score`; discard workspaces with zero score,
+/// and return a list of workspaces IDs with non-zero score, ordered by score in descending order.
+/// For workspaces with equal score the order as in storage.json is preserved.
 fn find_matching_workspaces<'a, I, S, T, P>(
     workspaces: I,
     terms: &'a [S],
-) -> impl Iterator<Item = T> + 'a
+) -> Vec<T>
 where
     I: Iterator<Item = (T, P)> + 'a,
     P: Borrow<RecentWorkspace>,
-    T: AsRef<str>,
     S: AsRef<str>,
 {
-    workspaces.filter_map(move |(id, project)| {
-        Some(id).filter(move |_| {
-            terms.iter().all(|term| {
-                project
-                    .borrow()
-                    .name
-                    .to_lowercase()
-                    .contains(&term.as_ref().to_lowercase())
-            }) || terms.iter().all(|term| {
-                project
-                    .borrow()
-                    .url
-                    .to_lowercase()
-                    .contains(&term.as_ref().to_lowercase())
-            })
+    let mut matches: Vec<(f64, T)> = workspaces
+        .filter_map(move |(id, workspace)| {
+            let score = match_score(workspace.borrow(), terms);
+            if 0.0 < score {
+                Some((score, id))
+            } else {
+                None
+            }
         })
-    })
+        .collect();
+    // Sort by score, descending
+    matches.sort_by(|(score_a, _), (score_b, _)| score_b.partial_cmp(score_a).unwrap());
+    matches.into_iter().map(move |(_, id)| id).collect()
 }
 
 /// A DBus search provider for a VSCode variant.
@@ -263,6 +283,7 @@ impl VscodeSearchProvider {
         })?;
 
         let ids = find_matching_workspaces(self.recent_workspaces.iter(), terms.as_slice())
+            .into_iter()
             .map(String::to_owned)
             .collect();
         debug!("Found ids {:?} for {}", ids, self.app.get_id().unwrap());
@@ -290,6 +311,7 @@ impl VscodeSearchProvider {
             .filter_map(|id| self.recent_workspaces.get(id).map(|p| (id, p)));
 
         let ids = find_matching_workspaces(candidates, terms.as_slice())
+            .into_iter()
             .map(String::to_owned)
             .collect();
         debug!("Found ids {:?} for {}", ids, self.app.get_id().unwrap());
@@ -551,7 +573,7 @@ mod tests {
         use crate::{find_matching_workspaces, RecentWorkspace};
 
         fn do_match<'a>(projects: &[(&'a str, RecentWorkspace)], terms: &[&str]) -> Vec<&'a str> {
-            find_matching_workspaces(projects.iter().map(|(s, p)| (*s, p)), terms).collect()
+            find_matching_workspaces(projects.iter().map(|(s, p)| (*s, p)), terms)
         }
 
         #[test]
@@ -616,6 +638,51 @@ mod tests {
                 },
             )];
             assert_eq!(do_match(&workspaces, &["Mdcat"]), ["foo"]);
+        }
+
+
+        #[test]
+        fn matches_in_name_rank_higher() {
+            let projects = vec![
+                (
+                    "1",
+                    RecentWorkspace {
+                        name: "bar".to_string(),
+                        // This matches foo as well because of /home/foo
+                        url: "file:///home/foo/dev/bar".to_string(),
+                    },
+                ),
+                (
+                    "2",
+                    RecentWorkspace {
+                        name: "foo".to_string(),
+                        url: "/home/foo/dev/foo".to_string(),
+                    },
+                ),
+            ];
+            assert_eq!(do_match(&projects, &["foo"]), ["2", "1"]);
+        }
+
+        #[test]
+        fn matches_at_end_of_url_rank_higher() {
+            let projects = vec![
+                (
+                    "1",
+                    RecentWorkspace {
+                        name: "p1".to_string(),
+                        // This matches foo as well because of /home/foo
+                        url: "file:///home/foo/dev/bar".to_string(),
+                    },
+                ),
+                (
+                    "2",
+                    RecentWorkspace {
+                        name: "p1".to_string(),
+                        url: "file:///home/foo/dev/foo".to_string(),
+                    },
+                ),
+            ];
+            assert_eq!(do_match(&projects, &["foo"]), ["2", "1"]);
         }
     }